feat: enrich join requests in repository
Move user profile lookup and merging logic for join requests
from DashboardTeamMembersFragment into a new internal method
`fetchEnrichedTeamJoinRequests` in DashboardTeamsRepository.
The fragment now cleanly consumes a list of ready-to-render
TeamJoinRequestUiModel objects.
